core/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Django==5.1.2
django-kafka-streamer==1.3.1
django-ninja==1.3.0
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
ipython==8.28.0
jedi==0.19.1
kafka-python-ng==2.2.3
kombu==5.4.2
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-binary==3.2.3
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
redis==5.1.1
sentry-sdk==2.16.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
vine==5.1.0
wcwidth==0.2.13
-e ../protocol
